fix linter

+++ internal/app/client/movie.go
@@ -14,6 +14,7 @@ import (
 //go:generate mockgen -source=movie.go -destination=../user/service/mocks/movie_mock.go
 //go:generate mockgen -source=movie.go -destination=../movie/service/mocks/mock.go
 
+// MovieClientInterface интерфейс клиента фильмов
 type MovieClientInterface interface {
 	// GetMovieByGenre(ctx context.Context, genre string) ([]models.MovieShortInfo, error)
 	GetMovie(ctx context.Context, mvID int) (*models.MovieInfo, error)
@@ -229,7 +230,7 @@ func (m MovieClient) GetActor(ctx context.Context, actorID int) (*models.ActorIn
 // var respp = make([]models.MovieShortInfo, 0, len(respMovie))
 
 // 	for i, movie := range respMovie {
-// 		respp[i].ID = int(movie.Id)
+// 		respp[i].ID = int(movie.ID)
 // 		respp[i].CardURL = movie.CardUrl
 // 		respp[i].MovieType = movie.MovieType
 // 		respp[i].AlbumURL = movie.AlbumUrl
@@ -333,6 +334,7 @@ func (m MovieClient) GetFavorites(ctx context.Context, mvIDs []uint64) ([]models
 	return ans, nil
 }
 
+// GetUserRating получение рейтинга
 func (m MovieClient) GetUserRating(ctx context.Context, movieID, userID int32) (int32, error) {
 	start := time.Now()
 	method := "GetUserRating"
@@ -351,14 +353,15 @@ func (m MovieClient) GetUserRating(ctx context.Context, movieID, userID int32) (
 	return int32(resp.Rating.Rating), nil
 }
 
+// AddOrUpdateRating добавление рейтинга
 func (m MovieClient) AddOrUpdateRating(ctx context.Context, movieID, userID, rating int32) error {
 	start := time.Now()
 	method := "AddOrUpdateRating"
 
 	_, err := m.movieMS.AddOrUpdateRating(ctx, &movie.AddOrUpdateRatingRequest{
-		MovieId: int32(movieID),
-		UserId:  int32(userID),
-		Rating:  int32(rating),
+		MovieId: movieID,
+		UserId:  userID,
+		Rating:  rating,
 	})
 
 	saveMetric(start, movieClient, method, err)
@@ -366,6 +369,7 @@ func (m MovieClient) AddOrUpdateRating(ctx context.Context, movieID, userID, rat
 	return err
 }
 
+// DeleteUserRating удаление рейтинга
 func (m *MovieClient) DeleteUserRating(ctx context.Context, userID, movieID int32) error {
 	start := time.Now()
 	method := "AddOrUpdateRating"
